# contratos/openapi.yaml
openapi: 3.1.0

info:
  title: API de Reservas CoLabora
  version: 1.0.0
  description: |
    API pública del sistema de Reservas de la Red CoLabora. Permite registrar usuarios,
    consultar la disponibilidad de salas y reservar puestos. La disponibilidad se consulta
    internamente al servicio de Espacios mediante gRPC.

    **Autenticación.** Bearer con JWT. El token se obtiene en `POST /v1/auth/login`
    y dura 60 minutos.

    **Errores.** Todas las respuestas de error usan Problem Details (RFC 9457,
    `application/problem+json`).

    **Enlaces (HATEOAS).** Los recursos incluyen `_links` inspirados en HAL, extendidos con
    `method` y, cuando la acción lleva un cuerpo fijo, `body`. Un enlace solo aparece si quien
    consulta puede ejecutar esa acción en el estado actual del recurso.

    **Evolución.** Los cambios compatibles se agregan a `/v1`; un cambio incompatible exige `/v2`.
    Los clientes deben ignorar los campos desconocidos de las respuestas. Los endpoints en retiro
    se anuncian con `deprecated: true` y los headers `Deprecation` y `Sunset`.

servers:
  - url: http://localhost:8000
    description: Entorno local levantado con docker compose


# Por defecto, todo endpoint exige token. Los públicos lo anulan con `security: []`.tags:
  - name: Autenticación
  - name: Usuarios
  - name: Salas
  - name: Reservas

security:
  - bearerAuth: []

paths: 
  # Autenticación
  /v1/auth/login:
    post:
      tags: [Autenticación]
      operationId: iniciarSesion
      summary: Obtener un token de acceso
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credenciales'
      responses:
        '200':
          description: Credenciales correctas.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '400':
          $ref: '#/components/responses/JsonMalFormado'
        '401':
          description: Email o contraseña incorrectos.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problema'
        '422':
          $ref: '#/components/responses/ValidacionFallida'
  
  # Usuarios
  /v1/usuarios:
    post:
      tags: [Usuarios]
      operationId: registrarUsuario
      summary: Registrar una cuenta
      description: |
        Registro abierto. Sin token, o con token de rol `usuario`, la cuenta se crea
        siempre con rol `usuario` y el campo `rol` se ignora. Solo un administrador puede crear
        cuentas con rol `administrador`.
      security:
        - {}
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsuarioNuevo'
      responses:
        '201':
          description: Cuenta creada.
          headers:
            Location:
              $ref: '#/components/headers/Location'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
        '400':
          $ref: '#/components/responses/JsonMalFormado'
        '401':
          $ref: '#/components/responses/NoAutenticado'
        '409':
          description: El email ya está registrado.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problema'
        '422':
          $ref: '#/components/responses/ValidacionFallida'

    get:
      tags: [Usuarios]
      operationId: listarUsuarios
      summary: Listar las cuentas (solo administrador)
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Página de usuarios.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListaUsuarios'
        '401':
          $ref: '#/components/responses/NoAutenticado'
        '403':
          $ref: '#/components/responses/SinPermiso'
        '422':
          $ref: '#/components/responses/ValidacionFallida'

  /v1/usuarios/me:
    get:
      tags: [Usuarios]
      operationId: obtenerUsuarioActual
      summary: Ver la cuenta del usuario autenticado
      responses:
        '200':
          description: Cuenta del usuario dueño del token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
        '401':
          $ref: '#/components/responses/NoAutenticado'

  /v1/usuarios/{usuario_id}:
    get:
      tags: [Usuarios]
      operationId: obtenerUsuario
      summary: Ver una cuenta
      description: Un usuario solo puede ver su propia cuenta; la de otro responde 404.
      parameters:
        - $ref: '#/components/parameters/UsuarioId'
      responses:
        '200':
          description: Cuenta encontrada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
        '401':
          $ref: '#/components/responses/NoAutenticado'
        '404':
          $ref: '#/components/responses/NoEncontrado'

  # ------------------------------------------------------------------ Salas

  /v1/salas:
    get:
      tags: [Salas]
      operationId: listarDisponibilidad
      summary: Salas con sus puestos libres en cada franja de una fecha
      description: |
        Consulta a Espacios (`ListarDisponibilidad`). El resultado puede venir de la caché
        mientras esté vigente. Si Espacios no responde y no hay dato vigente, responde 503 o 504.
      parameters:
        - $ref: '#/components/parameters/Fecha'
      responses:
        '200':
          description: Grilla de disponibilidad de la fecha.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GrillaDisponibilidad'
        '401':
          $ref: '#/components/responses/NoAutenticado'
        '422':
          $ref: '#/components/responses/ValidacionFallida'
        '503':
          $ref: '#/components/responses/EspaciosNoDisponible'
        '504':
          $ref: '#/components/responses/EspaciosSinRespuesta'

  /v1/salas/{sala_id}/disponibilidad:
    get:
      tags: [Salas]
      operationId: consultarDisponibilidad
      summary: Puestos libres de una sala en una franja
      description: Consulta a Espacios (`ConsultarDisponibilidad`), con caché mientras esté vigente.
      parameters:
        - $ref: '#/components/parameters/SalaId'
        - $ref: '#/components/parameters/Fecha'
        - $ref: '#/components/parameters/HoraInicio'
        - $ref: '#/components/parameters/HoraFin'
      responses:
        '200':
          description: Disponibilidad de la sala en la franja.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Disponibilidad'
        '401':
          $ref: '#/components/responses/NoAutenticado'
        '404':
          $ref: '#/components/responses/NoEncontrado'
        '422':
          description: La franja no corresponde a un bloque definido o los parámetros no son válidos.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problema'
        '503':
          $ref: '#/components/responses/EspaciosNoDisponible'
        '504':
          $ref: '#/components/responses/EspaciosSinRespuesta'

  # ------------------------------------------------------------------ Reservas

  /v1/reservas:
    post:
      tags: [Reservas]
      operationId: crearReserva
      summary: Reservar un puesto
      description: |
        Pide a Espacios ocupar un puesto (`OcuparPuesto`) en una sola operación atómica y, si lo
        logra, registra la reserva.

        - Un usuario solo reserva para sí mismo. Un administrador puede indicar `titular_id` para
          reservar a nombre de otro usuario.
        - Con `Idempotency-Key`, reintentar con la misma clave y el mismo cuerpo devuelve la
          respuesta original sin reservar de nuevo.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReservaNueva'
      responses:
        '201':
          description: Reserva creada; el puesto quedó ocupado en Espacios.
          headers:
            Location:
              $ref: '#/components/headers/Location'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reserva'
        '400':
          $ref: '#/components/responses/JsonMalFormado'
        '401':
          $ref: '#/components/responses/NoAutenticado'
        '403':
          description: Un usuario intentó reservar a nombre de otro.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problema'
        '409':
          description: |
            Conflicto. Según `type`:
            - `/problemas/sin-puestos`: la sala no tiene puestos libres en esa franja.
            - `/problemas/idempotencia-en-curso`: otra petición con la misma `Idempotency-Key`
              todavía se está procesando.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problema'
        '422':
          description: |
            Datos inválidos. Incluye: la sala o el titular no existen, la franja no corresponde a
            un bloque definido, o la `Idempotency-Key` ya se usó con un cuerpo distinto
            (`/problemas/idempotencia-reutilizada`).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problema'
        '500':
          description: |
            Espacios ocupó el puesto pero no se pudo registrar la reserva; el puesto se libera
            como compensación.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problema'
        '503':
          $ref: '#/components/responses/EspaciosNoDisponible'
        '504':
          $ref: '#/components/responses/EspaciosSinRespuesta'

    get:
      tags: [Reservas]
      operationId: listarReservas
      summary: Listar reservas
      description: Un usuario ve solo sus reservas; un administrador ve todas. No consulta a Espacios.
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Página de reservas.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListaReservas'
        '401':
          $ref: '#/components/responses/NoAutenticado'
        '422':
          $ref: '#/components/responses/ValidacionFallida'

  /v1/reservas/{reserva_id}:
    get:
      tags: [Reservas]
      operationId: obtenerReserva
      summary: Ver una reserva
      description: La reserva de otro usuario responde 404. No consulta a Espacios.
      parameters:
        - $ref: '#/components/parameters/ReservaId'
      responses:
        '200':
          description: Reserva encontrada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reserva'
        '401':
          $ref: '#/components/responses/NoAutenticado'
        '404':
          $ref: '#/components/responses/NoEncontrado'

    patch:
      tags: [Reservas]
      operationId: cancelarReserva
      summary: Cancelar una reserva
      description: |
        Único cambio admitido: `{"estado": "CANCELADA"}` (JSON Merge Patch, RFC 7396).

        La reserva queda cancelada de inmediato y la liberación del puesto en Espacios se
        encola; se procesa en segundos, o cuando Espacios vuelva si está caído. Por eso
        esta operación responde 200 aunque Espacios no esté disponible.

        Es idempotente: cancelar una reserva ya cancelada devuelve 200 con la reserva.
      parameters:
        - $ref: '#/components/parameters/ReservaId'
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: '#/components/schemas/CancelacionReserva'
      responses:
        '200':
          description: Reserva cancelada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reserva'
        '400':
          $ref: '#/components/responses/JsonMalFormado'
        '401':
          $ref: '#/components/responses/NoAutenticado'
        '404':
          $ref: '#/components/responses/NoEncontrado'
        '422':
          $ref: '#/components/responses/ValidacionFallida'

components:

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token obtenido en `POST /v1/auth/login`. Expira a los 60 minutos.

  # ------------------------------------------------------------------ Parámetros

  parameters:
    UsuarioId:
      name: usuario_id
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    SalaId:
      name: sala_id
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    ReservaId:
      name: reserva_id
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    Fecha:
      name: fecha
      in: query
      required: true
      description: Fecha en formato ISO 8601.
      schema:
        type: string
        format: date
        examples: ['2026-10-01']
    HoraInicio:
      name: hora_inicio
      in: query
      required: true
      schema:
        $ref: '#/components/schemas/Hora'
    HoraFin:
      name: hora_fin
      in: query
      required: true
      schema:
        $ref: '#/components/schemas/Hora'
    Limit:
      name: limit
      in: query
      required: false
      description: Cantidad máxima de elementos por página.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Offset:
      name: offset
      in: query
      required: false
      description: Cantidad de elementos a saltar desde el inicio.
      schema:
        type: integer
        minimum: 0
        default: 0
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        Clave única elegida por el cliente (se recomienda un UUID) para reintentar sin crear
        reservas duplicadas. Su alcance es por usuario.
      schema:
        type: string
        maxLength: 255
        examples: ['8e03978e-40d5-43e8-bc93-6894a57f9324']

  # ------------------------------------------------------------------ Headers

  headers:
    Location:
      description: URL del recurso creado.
      schema:
        type: string
        examples: ['/v1/reservas/7']
    RetryAfter:
      description: Segundos sugeridos antes de reintentar.
      schema:
        type: integer
        examples: [10]
    WwwAuthenticate:
      description: Esquema de autenticación esperado (RFC 6750).
      schema:
        type: string
        examples: ['Bearer']

  # Respuestas reutilizables

  responses:
    JsonMalFormado:
      description: El cuerpo no es JSON válido.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problema'
    ValidacionFallida:
      description: Los datos no cumplen el formato esperado.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problema'
    NoAutenticado:
      description: Falta el token, no es válido o expiró.
      headers:
        WWW-Authenticate:
          $ref: '#/components/headers/WwwAuthenticate'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problema'
    SinPermiso:
      description: El usuario está autenticado, pero su rol no permite esta operación.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problema'
    NoEncontrado:
      description: El recurso no existe o no pertenece al usuario.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problema'
    EspaciosNoDisponible:
      description: El servicio de Espacios no está disponible. No se realizó ninguna operación.
      headers:
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problema'
    EspaciosSinRespuesta:
      description: |
        El servicio de Espacios no respondió a tiempo. El resultado es incierto; si era una
        reserva, se compensa liberando el puesto.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problema'

  # Esquemas

  schemas:
    Hora:
      type: string
      pattern: '^([01][0-9]|2[0-3]):[0-5][0-9]$'
      description: Hora de reloj en formato HH:MM, 24 horas.
      examples: ['09:00']
    Problema:
      type: object
      description: Error en formato Problem Details (RFC 9457).
      required: [type, title, status]
      properties:
        type:
          type: string
          description: URI que identifica el tipo de problema.
          examples: ['/problemas/sin-puestos']
        title:
          type: string
          examples: ['Sin puestos disponibles']
        status:
          type: integer
          examples: [409]
        detail:
          type: string
          examples: ['La sala 3 no tiene puestos libres el 2026-10-01 de 09:00 a 10:00.']
        instance:
          type: string
          examples: ['/v1/reservas']
        errores:
          type: array
          description: Extensión para errores de validación; un elemento por campo inválido.
          items:
            type: object
            required: [campo, mensaje]
            properties:
              campo:
                type: string
                examples: ['email']
              mensaje:
                type: string
                examples: ['No es un email válido.']
    Enlace:
      type: object
      description: |
        Acción disponible sobre el recurso. `body`, si existe, es el cuerpo que hay
        que enviar.
      required: [href, method]
      properties:
        href:
          type: string
          examples: ['/v1/reservas/7']
        method:
          type: string
          enum: [GET, POST, PATCH]
        body:
          type: object
          examples: [{estado: CANCELADA}]
    EnlacesPagina:
      type: object
      required: [self]
      properties:
        self:
          $ref: '#/components/schemas/Enlace'
        siguiente:
          $ref: '#/components/schemas/Enlace'
        anterior:
          $ref: '#/components/schemas/Enlace'

    # Autenticación
    Credenciales:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    Token:
      type: object
      required: [access_token, token_type, expires_in]
      properties:
        access_token:
          type: string
        token_type:
          type: string
          const: Bearer
        expires_in:
          type: integer
          description: Segundos de validez del token.
          examples: [3600]
    
    # Usuarios
    Rol:
      type: string
      enum: [usuario, administrador]
    UsuarioNuevo:
      type: object
      required: [nombre, email, password]
      properties:
        nombre:
          type: string
          minLength: 1
          maxLength: 100
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          maxLength: 72
          description: |
            Entre 8 caracteres y 72 bytes en UTF-8 (límite de bcrypt). Un carácter
            como "ñ" ocupa 2 bytes, por lo que el máximo en caracteres puede ser menor que 72.
        rol:
          $ref: '#/components/schemas/Rol'
          description: Solo lo considera un administrador; en otro caso se ignora.
    Usuario:
      type: object
      required: [id, nombre, email, rol, creado_en, _links]
      properties:
        id:
          type: integer
          format: int64
        nombre:
          type: string
        email:
          type: string
          format: email
        rol:
          $ref: '#/components/schemas/Rol'
        creado_en:
          type: string
          format: date-time
        _links:
          type: object
          required: [self]
          properties:
            self:
              $ref: '#/components/schemas/Enlace'
    ListaUsuarios:
      type: object
      required: [items, total, limit, offset, _links]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Usuario'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        _links:
          $ref: '#/components/schemas/EnlacesPagina'

    # Salas y disponibilidad 
    Sala:
      type: object
      required: [id, nombre, capacidad]
      properties:
        id:
          type: integer
          format: int64
        nombre:
          type: string
        capacidad:
          type: integer
          description: Total de puestos de la sala.
    FranjaDisponible:
      type: object
      required: [hora_inicio, hora_fin, puestos_libres, _links]
      properties:
        hora_inicio:
          $ref: '#/components/schemas/Hora'
        hora_fin:
          $ref: '#/components/schemas/Hora'
        puestos_libres:
          type: integer
        _links:
          type: object
          description: '`reservar` aparece solo si quedan puestos libres.'
          properties:
            reservar:
              $ref: '#/components/schemas/Enlace'
    SalaConDisponibilidad:
      allOf:
        - $ref: '#/components/schemas/Sala'
        - type: object
          required: [franjas]
          properties:
            franjas:
              type: array
              items:
                $ref: '#/components/schemas/FranjaDisponible'
    GrillaDisponibilidad:
      type: object
      required: [fecha, salas]
      properties:
        fecha:
          type: string
          format: date
        salas:
          type: array
          items:
            $ref: '#/components/schemas/SalaConDisponibilidad'
    Disponibilidad:
      type: object
      required: [sala, fecha, hora_inicio, hora_fin, puestos_libres, _links]
      properties:
        sala:
          $ref: '#/components/schemas/Sala'
        fecha:
          type: string
          format: date
        hora_inicio:
          $ref: '#/components/schemas/Hora'
        hora_fin:
          $ref: '#/components/schemas/Hora'
        puestos_libres:
          type: integer
        _links:
          type: object
          required: [self]
          properties:
            self:
              $ref: '#/components/schemas/Enlace'
            reservar:
              $ref: '#/components/schemas/Enlace'

    # --- Reservas ---
    EstadoReserva:
      type: string
      enum: [ACTIVA, CANCELADA]
    ReservaNueva:
      type: object
      required: [sala_id, fecha, hora_inicio, hora_fin]
      properties:
        sala_id:
          type: integer
          format: int64
        fecha:
          type: string
          format: date
        hora_inicio:
          $ref: '#/components/schemas/Hora'
        hora_fin:
          $ref: '#/components/schemas/Hora'
        titular_id:
          type: integer
          format: int64
          description: |
            Usuario para quien es la reserva. Solo un administrador puede indicar a otro usuario;
            si se omite, el titular es quien reserva.
    Reserva:
      type: object
      required:
        - id
        - estado
        - sala_id
        - sala_nombre
        - fecha
        - hora_inicio
        - hora_fin
        - titular_id
        - creada_por_id
        - creada_en
        - _links
      properties:
        id:
          type: integer
          format: int64
        estado:
          $ref: '#/components/schemas/EstadoReserva'
        sala_id:
          type: integer
          format: int64
        sala_nombre:
          type: string
          description: Nombre de la sala al momento de reservar (copia intencional).
        fecha:
          type: string
          format: date
        hora_inicio:
          $ref: '#/components/schemas/Hora'
        hora_fin:
          $ref: '#/components/schemas/Hora'
        titular_id:
          type: integer
          format: int64
          description: Usuario para quien es la reserva.
        creada_por_id:
          type: integer
          format: int64
          description: Usuario que realizó la reserva (tomado del token).
        creada_en:
          type: string
          format: date-time
        cancelada_en:
          type: [string, 'null']
          format: date-time
        _links:
          type: object
          required: [self, sala]
          description: '`cancelar` aparece solo si la reserva está ACTIVA y quien consulta es el titular o un administrador.'
          properties:
            self:
              $ref: '#/components/schemas/Enlace'
            sala:
              $ref: '#/components/schemas/Enlace'
            cancelar:
              $ref: '#/components/schemas/Enlace'
    CancelacionReserva:
      type: object
      description: JSON Merge Patch (RFC 7396). Solo se admite cancelar.
      required: [estado]
      properties:
        estado:
          type: string
          const: CANCELADA
    ListaReservas:
      type: object
      required: [items, total, limit, offset, _links]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Reserva'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        _links:
          $ref: '#/components/schemas/EnlacesPagina'